Proper tuning

+++ sintab.h
@@ -2,6 +2,6 @@
 #define sintab_h
 
 #define SINTAB_LEN 256
-extern int16_t sintab[SINTAB_LEN];
+extern int8_t sintab[SINTAB_LEN];
 
 #endif
